requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
streamlit>=1.28.0
plotly>=5.17.0
openai>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
reportlab>=4.0.0
fpdf2>=2.7.0
pytest>=7.4.0
google-generativeai>=0.3.0
gunicorn>=21.0.0